:recycle: delegate components
